tf_requirements.txt
absl-py==1.4.0
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.5.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
docker-pycreds==0.4.0
dp-accounting==0.4.3
etils==1.5.2
exceptiongroup==1.1.3
executing==2.0.1
farmhashpy==0.4.0
fastjsonschema==2.18.1
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
fsspec==2023.10.0
gast==0.5.4
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-vizier==0.1.11
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
grpcio-tools==1.59.2
h5py==3.10.0
httplib2==0.22.0
idna==3.4
immutabledict==2.2.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.26.0
ipython==8.17.2
isoduration==20.11.0
jax==0.4.14
jaxlib==0.4.14
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kaggle==1.5.16
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.25.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.16.5
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
overrides==7.4.0
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
portalocker==2.8.2
portpicker==1.6.0
prometheus-client==0.18.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
sacrebleu==2.3.1
scikit-learn==1.3.2
scipy==1.9.3
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.35.0
seqeval==1.2.2
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.20
stack-data==0.6.3
statsmodels==0.14.0
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-compression==2.14.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.14.0
tensorflow-federated==0.64.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.13.1
tensorflow-model-optimization==0.7.5
tensorflow-privacy==0.8.12
tensorflow-probability==0.22.1
tensorflow-text==2.14.0
tensorrt==8.5.3.1
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
tf-models-official==2.14.2
tf-slim==1.1.0
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.5.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.7
wandb==0.16.0
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
ipfshttpclient==0.7.0